Move DateTimePicker to the pickers folder
