--- src/DateTimePicker/stories/overview.jsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-import ComponentOverview from '../../ComponentOverview'
-import { DateTimePicker } from '../../index'
-
-export default () => <ComponentOverview component={DateTimePicker} />
