+++ src/index.js
@@ -23,7 +23,7 @@ export { ThemeProvider, themes } from './styles'
 export { default as Button } from './Button'
 export { default as Card } from './card/Card'
 export { default as CardAction } from './card/CardAction'
-export { default as DateTimePicker } from './DateTimePicker'
+export { default as DateTimePicker } from './pickers/DateTimePicker'
 export { default as Dialog } from './dialog/Dialog'
 export { default as DialogAction } from './dialog/DialogAction'
 export { default as DialogActions } from './dialog/DialogActions'
